remake test for certificate compliance check and validation, sq fixes
