+++ src/components/_pages/auditLogs/index.tsx
@@ -85,8 +85,6 @@ function AuditLogs() {
 
     const onInfoClick = useCallback(
         (log: AuditLogDto) => {
-            console.log('log', log);
-
             dispatch(
                 userInterfaceActions.showGlobalModal({
                     content: (
